blockchain_python_2/backend/requirements.txt
fastapi
requests
uvicorn
python-dotenv
strawberry-graphql[fastapi]
autopep8